refactor(authorization): delegate authentication checks to the authentication plugin
The authorization plugin's require_authentication check and its cache-key
scope extraction both read authentication's raw JWT-claims context key
directly, which an unresolved review comment (XXX(@goto-bus-stop)) flagged
as odd. Add has_authenticated_jwt() and jwt_scopes() to the authentication
plugin so authorization asks a clean API instead of reaching into another
plugin's context key itself. No behavior or config change.

Fixes ROUTER-985
