+++ apollo-router/src/plugins/authorization/mod.rs
@@ -34,7 +34,7 @@ use crate::json_ext::Path;
 use crate::layers::ServiceBuilderExt;
 use crate::plugin::Plugin;
 use crate::plugin::PluginInit;
-use crate::plugins::authentication::APOLLO_AUTHENTICATION_JWT_CLAIMS;
+use crate::plugins::authentication;
 use crate::query_planner::FilteredQuery;
 use crate::query_planner::QueryKey;
 use crate::services::execution;
@@ -300,18 +300,9 @@ impl AuthorizationPlugin {
     }
 
     pub(crate) fn update_cache_key(context: &Context) {
-        let is_authenticated = context.contains_key(APOLLO_AUTHENTICATION_JWT_CLAIMS);
+        let is_authenticated = authentication::has_authenticated_jwt(context);
 
-        let request_scopes = context
-            .get_json_value(APOLLO_AUTHENTICATION_JWT_CLAIMS)
-            .and_then(|value| {
-                value.as_object().and_then(|object| {
-                    object.get("scope").and_then(|v| {
-                        v.as_str()
-                            .map(|s| s.split(' ').map(|s| s.to_string()).collect::<HashSet<_>>())
-                    })
-                })
-            });
+        let request_scopes = authentication::jwt_scopes(context);
         let query_scopes = context.get_json_value(REQUIRED_SCOPES_KEY).and_then(|v| {
             v.as_array().map(|v| {
                 v.iter()
@@ -590,12 +581,10 @@ impl Plugin for AuthorizationPlugin {
         if self.require_authentication {
             ServiceBuilder::new()
                 .checkpoint(move |request: supergraph::Request| {
-                    // XXX(@goto-bus-stop): Why are we doing this here, as opposed to the
-                    // authentication plugin, which manages this context value?
-                    if request
-                        .context
-                        .contains_key(APOLLO_AUTHENTICATION_JWT_CLAIMS)
-                    {
+                    // Whether to reject unauthenticated requests is an authorization policy,
+                    // same as `@authenticated`/`@requiresScopes`/`@policy` — authentication only
+                    // verifies tokens, it doesn't decide this.
+                    if authentication::has_authenticated_jwt(&request.context) {
                         Ok(ControlFlow::Continue(request))
                     } else {
                         tracing::error!("rejecting unauthenticated request");
